+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog ##
 #### dev ####
+* Make workshop support still work on older versions of Slay the Spire
 
 #### v3.25.1 ####
 * Fix steamworks4j version for Slay the Spire v2.3